+++ src/main/java/utils/HttpResponseUtils.java
@@ -1,0 +1,2 @@
+package utils;public class HttpResponseUtils {
+}
