+++ src/main/java/view/impl/RedirectResponseView.java
@@ -1,0 +1,15 @@
+package view.impl;
+
+import view.View;
+
+public class RedirectResponseView extends View {
+
+	public RedirectResponseView(String view) {
+		setBody(render(view));
+	}
+
+	@Override
+	public byte[] render(String view) {
+		return new byte[0];
+	}
+}
